style: RenderOrderInput 스타일 적용
